+++ ft_puthexa.c
@@ -6,13 +6,13 @@
 /*   By: mle-biha <mle-biha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/17 14:23:12 by mle-biha          #+#    #+#             */
-/*   Updated: 2022/08/18 21:01:46 by mle-biha         ###   ########.fr       */
+/*   Updated: 2022/08/19 15:15:06 by mle-biha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_puthexa(int nb, char *base)
+int	ft_puthexa(size_t nb, char *base)
 {
 	int	i;
